requirements.txt
Flask-SQLAlchemy==2.4.4
Flask==1.1.2
alembic==1.4.3
bandit==1.7.0
black==20.8b1
configly[yaml]==0.2.0
flake8==3.8.4
flask-mongoengine==0.9.0
flask-redis==0.4.0
isort==5.6.4
marshmallow==3.9.1
mypy-extensions==0.4.3
mypy==0.790
pendulum==2.1.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
